cmd: move analyze command to its own file
